Hypothetical worlds and config based values
Reworks a great number of things, this is not going to be an adaquate commit message. The big ticket items are improvements to allow hypothetical worlds branched off of the main world (useful for animating changes on top of the main world state as well as for determining likely results of actions) and improvements to loading information from configuration.